flask-server-panel/requirements.txt
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask==2.3.3
Flask-Cors==4.0.0
waitress==2.1.2
